:recycle: Extract LinkOptions into own component
