+++ src/components/wysiwygEditor/WysiwygEditor.js
@@ -1,4 +1,4 @@
-import React, { useState, useRef, useEffect } from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { Editor } from 'react-draft-wysiwyg';
 import cx from 'classnames';
@@ -11,103 +11,12 @@ import {
 } from '@teamleader/ui-icons';
 
 import Box from '../box';
-import IconButton from '../iconButton';
 import ValidationText from '../validationText';
-import Popover from '../popover';
-import Input from '../input';
-import Button from '../button';
-import Label from '../label';
 
 import InlineStylingOptions from './InlineStylingOptions';
 import ListStylingOptions from './ListStylingOptions';
+import LinkOptions from './LinkOptions';
 import theme from './theme.css';
-
-const LinkOptions = ({
-  config: {
-    options,
-    link: { icon: linkIcon },
-    defaultTargetOption,
-  },
-  currentState,
-  currentState: { link, selectionText },
-  onChange,
-}) => {
-  const [isPopoverShown, setIsPopoverShown] = useState(false);
-  const iconButtonRef = useRef();
-  const [textValue, setTextValue] = useState(link?.title || selectionText);
-  const [linkValue, setLinkValue] = useState(link?.target);
-
-  useEffect(() => {
-    setTextValue(link?.title || selectionText);
-    setLinkValue(link?.target);
-  }, [link?.target, selectionText]);
-
-  const handleOpenPopoverClick = () => {
-    setIsPopoverShown(true);
-  };
-
-  const handleClosePopoverClick = () => {
-    setIsPopoverShown(false);
-  };
-
-  const onTextChange = ({ currentTarget: { value: newText } }) => {
-    setTextValue(newText);
-  };
-
-  const onLinkChange = ({ currentTarget: { value: newLink } }) => {
-    setLinkValue(newLink);
-  };
-
-  const handleAddLinkClick = () => {
-    onChange('link', textValue, linkValue, defaultTargetOption);
-    setIsPopoverShown(false);
-  };
-
-  return (
-    <>
-      <IconButton
-        icon={linkIcon}
-        color="black"
-        key="link"
-        title="link"
-        onClick={handleOpenPopoverClick}
-        selected={!!link || isPopoverShown}
-        marginRight={2}
-        ref={iconButtonRef}
-      />
-      <Popover
-        anchorEl={iconButtonRef?.current?.buttonNode?.boxNode?.current}
-        active={isPopoverShown}
-        backdrop="transparent"
-        onEscKeyDown={handleClosePopoverClick}
-        onOverlayClick={handleClosePopoverClick}
-        minWidth={276}
-        position="end"
-      >
-        <Box display="flex" flexDirection="column" padding={4}>
-          <Label htmlFor="linkText" helpText="">
-            Text
-            <Input id="linkText" autoFocus value={textValue || ''} onChange={onTextChange} />
-          </Label>
-          <Label htmlFor="link" helpText="" marginBottom={4}>
-            Link
-            <Input id="link" value={linkValue || ''} onChange={onLinkChange} placeholder="e.g. www.teamleader.eu" />
-          </Label>
-          <Box display="flex" justifyContent="flex-end" alignItems="center" marginTop={4}>
-            <Button level="secondary" label="Cancel" size="small" marginRight={3} onClick={handleClosePopoverClick} />
-            <Button
-              level="primary"
-              label="Add link"
-              size="small"
-              disabled={!linkValue || !textValue}
-              onClick={handleAddLinkClick}
-            />
-          </Box>
-        </Box>
-      </Popover>
-    </>
-  );
-};
 
 const toolbar = {
   options: ['inline', 'list', 'link'],
